fix: BigDecimal 수정에 맞춰 기존 테스트 코드 수정
- ci 실패로 인해 발생한 테스트 코드 확인 후 BigDecimal 변경에 맞춰서 코드 수정

+++ payment/src/test/java/com/example/payment/application/service/OrderPaymentApiServiceTest.java
@@ -69,8 +69,8 @@ class OrderPaymentApiServiceTest {
                         orderId,
                         UUID.randomUUID(),
                         List.of(UUID.randomUUID()),
-                        12000L,
-                        3000L
+                        BigDecimal.valueOf(12000L),
+                        BigDecimal.valueOf(3000L)
                 ));
 
         OrderPaymentApiResponse response = orderPaymentApiService.payOrder(request);
@@ -87,7 +87,7 @@ class OrderPaymentApiServiceTest {
                 .containsExactly(new OrderPaymentLineCommand(
                         request.orderLines().get(0).orderItemId(),
                         sellerId,
-                        12000L
+                        BigDecimal.valueOf(12000L)
                 ));
 
         ArgumentCaptor<OrderPaymentResultMessage> eventCaptor = ArgumentCaptor.forClass(OrderPaymentResultMessage.class);
